Fixed UpdateProjections, error in float cast. Removed from main.cpp

+++ camera.cc
@@ -87,8 +87,7 @@ void Camera::Movement(float delta_time, const std::vector<bool> &is_key_pressed_
 
 // Updates all the currently loaded shaders' projection uniforms
 void Camera::UpdateProjections() {
-  projection_matrix_ = glm::perspective(
-      fov_, float(width_ / height_), 0.1f, 100.0f);
+  projection_matrix_ = glm::perspective(fov_, (float)width_ / (float)height_, 0.1f, 100.0f);
   ShadersProjectionIterator iter(shaders_);
   iter = iter.begin();
   while (iter != iter.end()) {
